Move custom serde serializer to its own module
