+++ relayer/src/config.rs
@@ -1,5 +1,6 @@
 //! Relayer configuration
 
+mod proof_specs;
 pub mod reload;
 pub mod types;
 
@@ -395,7 +396,7 @@ pub struct ChainConfig {
     pub trusting_period: Option<Duration>,
     #[serde(default)]
     pub memo_prefix: Memo,
-    #[serde(default, with = "proof_specs_serde")]
+    #[serde(default, with = "self::proof_specs")]
     pub proof_specs: ProofSpecs,
 
     // these two need to be last otherwise we run into `ValueAfterTable` error when serializing to TOML
@@ -406,42 +407,6 @@ pub struct ChainConfig {
     pub packet_filter: PacketFilter,
     #[serde(default)]
     pub address_type: AddressType,
-}
-
-mod proof_specs_serde {
-    use core::fmt;
-    use ibc::core::ics23_commitment::specs::ProofSpecs;
-    use serde::{de, ser, Deserializer, Serializer};
-
-    pub fn serialize<S: Serializer>(
-        proof_specs: &ProofSpecs,
-        serializer: S,
-    ) -> Result<S::Ok, S::Error> {
-        let json_str = serde_json::to_string_pretty(proof_specs).map_err(ser::Error::custom)?;
-        serializer.serialize_str(&json_str)
-    }
-
-    struct ProofSpecsVisitor;
-
-    impl<'de> de::Visitor<'de> for ProofSpecsVisitor {
-        type Value = ProofSpecs;
-
-        fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-            formatter.write_str("ICS23 proof-specs serialized as a JSON array")
-        }
-
-        fn visit_str<E: de::Error>(self, v: &str) -> Result<Self::Value, E> {
-            serde_json::from_str(v).map_err(E::custom)
-        }
-
-        fn visit_string<E: de::Error>(self, v: String) -> Result<Self::Value, E> {
-            self.visit_str(&v)
-        }
-    }
-
-    pub fn deserialize<'de, D: Deserializer<'de>>(deserializer: D) -> Result<ProofSpecs, D::Error> {
-        deserializer.deserialize_string(ProofSpecsVisitor)
-    }
 }
 
 /// Attempt to load and parse the TOML config file as a `Config`.
